Crate node_archive

This crate provides types for fetching and unpacking Node distribution archives, which is a tarball for Unixes and a zipfile for Windows.

These docs show the top-level exports of this crate as re-exported of the tarball module (due to limitations of rustdoc); the top-level exports are re-exported from tarball on Unix operating systems and from zip on Windows operating systems.

Modules

tarball 	 [
Unix
]
Provides types and functions for fetching and unpacking a Node installation tarball in Unix operating systems.

zip 	 [
Windows
]
Provides types and functions for fetching and unpacking a Node installation zip file in Windows operating systems.

Structs

HttpError

Tarball
A Node installation tarball.

Traits

Archive

Functions

fetch
Fetch a remote Node archive from the given URL and cache its results at the given file path.

load
Load a cached Node archive from the specified file.