Module node_archive::tarball

This is supported on Unix only.

Provides types and functions for fetching and unpacking a Node installation tarball in Unix operating systems.

Structs

ByteRangesNotAcceptedError 	 [
Unix
]

MissingHeaderError 	 [
Unix
]

Tarball 	 [
Unix
]
A Node installation tarball.

UnexpectedContentLengthError 	 [
Unix
]

Functions

content_length 	 [
Unix
]
Determines the length of an HTTP response's content in bytes, using the HTTP "Content-Length" header.

fetch_isize 	 [
Unix
]
Fetches just the isize field (the field that indicates the uncompressed size) of a gzip file from a URL. This makes two round-trips to the server but avoids downloading the entire gzip file. For very small files it's unlikely to be more efficient than simply downloading the entire file up front.

fetch_uncompressed_size 	 [
Unix
]
Determines the uncompressed size of a gzip file hosted at the specified URL by fetching just the metadata associated with the file. This makes two round-trips to the server, so it is only more efficient than simply downloading the file if the file is large enough that downloading it is slower than the extra round trips.

headers_only 	 [
Unix
]
Fetches just the headers of a URL.

load_isize 	 [
Unix
]
Loads the isize field (the field that indicates the uncompressed size) of a gzip file from disk.

load_uncompressed_size 	 [
Unix
]
Determines the uncompressed size of the specified gzip file on disk.

unpack_isize 	 [
Unix
]
Unpacks the isize field from a gzip payload as a 64-bit integer.