Struct node_archive::tarball::Tarball

pub struct Tarball<S: Read> {
    compressed_size: u64,
    uncompressed_size: u64,
    data: S,
}

A Node installation tarball.

Fields
compressed_size: u64 uncompressed_size: u64 data: S
Methods

impl Tarball<File>

pub fn load(source: File) -> Result<Self, Error>

Loads a cached Node tarball from the specified file.

impl Tarball<TeeReader<Response, File>>

pub fn fetch(url: &str, cache_file: &Path) -> Result<Self, Error>

Initiate fetching of a Node tarball from the given URL, returning a tarball that can be streamed (and that tees its data to a cache file as it streams).