Struct node_archive::zip::Remote

pub struct Remote {
    cached: Cached,
}

A data source for fetching a Node zip archive from a remote server.

Fields
cached: Cached
Methods

impl Remote

pub fn fetch(url: &str, cache_file: &Path) -> Result<Remote, Error>

Initiate fetching of a Node zip archive from the given URL, returning a Remote data source.