Struct node_archive::zip::Zip

pub struct Zip<S: Read + Seek> {
    compressed_size: u64,
    data: S,
}

Fields
compressed_size: u64 data: S
Methods

impl Zip<File>

pub fn load(source: File) -> Result<Self>

Loads a cached Node zip archive from the specified file.

pub fn fetch(url: &str, cache_file: &Path) -> Result<Self, Error>

Initiate fetching of a Node zip archive from the given URL, returning a Remote data source.