Module node_archive::tarball

This is supported on Unix only.

Provides types and functions for fetching and unpacking a Node installation tarball in Unix operating systems.

Re-exports

use std::io::Read;

use std::io::Seek;

use std::io::SeekFrom;

use std::path::Path;

use std::fs::File;

use flate2::read::GzDecoder;

use reqwest::header::AcceptRanges;

use reqwest::header::ContentLength;

use reqwest::header::Range;

use reqwest::header::RangeUnit;

use reqwest::header::ByteRangeSpec;

use reqwest::Response;

use reqwest;

use tar;

use tee::TeeReader;

use progress_read::ProgressRead;

use failure;

use super::Archive;

Structs

ByteRangesNotAcceptedError

MissingHeaderError

Tarball
A Node installation tarball.

UnexpectedContentLengthError

Functions

content_length
Determines the length of an HTTP response's content in bytes, using the HTTP "Content-Length" header.

fetch_isize
Fetches just the isize field (the field that indicates the uncompressed size) of a gzip file from a URL. This makes two round-trips to the server but avoids downloading the entire gzip file. For very small files it's unlikely to be more efficient than simply downloading the entire file up front.

fetch_uncompressed_size
Determines the uncompressed size of a gzip file hosted at the specified URL by fetching just the metadata associated with the file. This makes two round-trips to the server, so it is only more efficient than simply downloading the file if the file is large enough that downloading it is slower than the extra round trips.

headers_only
Fetches just the headers of a URL.

load_isize
Loads the isize field (the field that indicates the uncompressed size) of a gzip file from disk.

load_uncompressed_size
Determines the uncompressed size of the specified gzip file on disk.

unpack_isize
Unpacks the isize field from a gzip payload as a 64-bit integer.